Clean up iterator implemenation a bit
